+++ database/factories/CampaignFactory.php
@@ -2,11 +2,13 @@
 
 namespace Database\Factories;
 
+use App\Models\Campaign;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Support\Str;
 
 /**
- * @extends \Illuminate\Database\Eloquent\Factories\Factory<\App\Models\Campaign>
+ * @extends Factory<Campaign>
  */
 class CampaignFactory extends Factory
 {
@@ -20,6 +22,7 @@ class CampaignFactory extends Factory
         return [
             'user_id' => User::factory(),
             'name' => fake()->sentence(3),
+            'slug' => Str::slug(fake()->unique()->words(4, true)),
             'description' => fake()->optional()->paragraph(),
         ];
     }